requirements.txt
# Core data science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=1.7.0
catboost>=1.2.0
optuna>=3.3.0

# Time series
statsmodels>=0.14.0
prophet>=1.1.4

# Explainability
shap>=0.43.0

# Vietnamese holidays
holidays>=0.35

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
